Changed all BigDecimal Values to ints (#45)

+++ src/main/java/monopoly/io/StreetImporter.java
@@ -20,20 +20,20 @@ public class StreetImporter extends CsvImporter<Street> {
 		}
 		Street street = new Street();
 		street.setName(line.getValue(0));
-		street.setPrice(line.getBigDecimalValue(1));
+		street.setPrice(line.getIntegerValue(1));
 		StreetDetails rentDetails = new StreetDetails();
 		rentDetails.setBaseRent(new Integer(line.getValue(2)));
-		rentDetails.setFirstHouseRent(line.getBigDecimalValue(3));
-		rentDetails.setSecondHouseRent(line.getBigDecimalValue(4));
-		rentDetails.setThirdHouseRent(line.getBigDecimalValue(5));
-		rentDetails.setFourthHouseRent(line.getBigDecimalValue(6));
-		rentDetails.setHotelRent(line.getBigDecimalValue(7));
-		rentDetails.setPricePerHouse(line.getBigDecimalValue(8));
-		rentDetails.setMortage(line.getBigDecimalValue(9));
+		rentDetails.setFirstHouseRent(line.getIntegerValue(3));
+		rentDetails.setSecondHouseRent(line.getIntegerValue(4));
+		rentDetails.setThirdHouseRent(line.getIntegerValue(5));
+		rentDetails.setFourthHouseRent(line.getIntegerValue(6));
+		rentDetails.setHotelRent(line.getIntegerValue(7));
+		rentDetails.setPricePerHouse(line.getIntegerValue(8));
+		rentDetails.setMortage(line.getIntegerValue(9));
 		street.setRentDetails(rentDetails);
 		StreetGroup group = streetGroups.getOrDefault(line.getValue(10), new StreetGroup());
 		group.getStreets().add(street);
-		group.setPricePerHouse(line.getBigDecimalValue(8));
+		group.setPricePerHouse(line.getIntegerValue(8));
 		street.setGroup(group);
 		return street;
 	}
